Split configuration
