+++ html/pfappserver/root/static.alt/src/globals/pfConfigurationPortalModules.js
@@ -1,0 +1,9 @@
+export const pfConfigurationPortalModuleViewFields = (context = {}) => {
+  return []
+}
+
+export const pfConfigurationPortalModuleViewDefaults = (context = {}) => {
+  return {
+    id: null
+  }
+}
